完成Proto的编写

+++ logic/Logic.Constant/Constant.cs
@@ -11,7 +11,7 @@ namespace Logic.Constant
     public static class Constant
     {
         public const double MoveSpeed = 5;
-        public const int FrameRate = 20;
+        public const double FrameRate = 20;
         public const double TimeInterval = 1 / FrameRate;
         public const double MoveDistancePerFrame = MoveSpeed / FrameRate;
         public const char messageSpiltSeperation = ',';
